Separated the kittens adn game controllers.
