+++ server/controllers/game/index.js
@@ -1,7 +1,5 @@
 const character = require('./character')
-const kitten = require('./kitten')
 
 module.exports = {
   character,
-  kitten,
 }
