+++ server/controllers/kittens/index.js
@@ -1,0 +1,5 @@
+const kitten = require('./kitten')
+
+module.exports = {
+  kitten,
+}
